+++ src/main/java/com/github/wintersteve25/tau/components/base/UIComponent.java
@@ -3,6 +3,7 @@ package com.github.wintersteve25.tau.components.base;
 import com.github.wintersteve25.tau.theme.Theme;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.resources.sounds.SimpleSoundInstance;
+import net.minecraft.client.sounds.SoundManager;
 import net.minecraft.sounds.SoundEvent;
 import com.github.wintersteve25.tau.layout.Layout;
 
@@ -10,8 +11,10 @@ import com.github.wintersteve25.tau.layout.Layout;
  * The base of all UI Components
  */
 public interface UIComponent {
+    SoundManager SOUND_MANAGER = Minecraft.getInstance().getSoundManager();
+
     UIComponent build(Layout layout, Theme theme);
-    
+
     default void playSound(SoundEvent soundEvent) {
         playSound(soundEvent, 0.25f);
     }
@@ -19,8 +22,8 @@ public interface UIComponent {
     default void playSound(SoundEvent soundEvent, float volume) {
         playSound(soundEvent, volume, 1);
     }
-    
+
     default void playSound(SoundEvent sound, float volume, float pitch) {
-        Minecraft.getInstance().getSoundManager().play(SimpleSoundInstance.forUI(sound, pitch, volume));
+        SOUND_MANAGER.play(SimpleSoundInstance.forUI(sound, pitch, volume));
     }
 }