+++ src/main/java/com/github/wintersteve25/tau/menu/TauContainerUI.java
@@ -5,7 +5,6 @@ import com.github.wintersteve25.tau.build.UIBuilder;
 import com.github.wintersteve25.tau.components.base.DynamicUIComponent;
 import com.github.wintersteve25.tau.layout.Layout;
 import com.github.wintersteve25.tau.theme.Theme;
-import com.github.wintersteve25.tau.utils.SimpleVec2i;
 import com.mojang.blaze3d.vertex.PoseStack;
 import net.minecraft.client.gui.GuiGraphics;
 import net.minecraft.client.gui.components.Renderable;
@@ -20,7 +19,7 @@ import java.util.List;
 
 public class TauContainerUI extends AbstractContainerScreen<TauContainerMenu> implements MenuAccess<TauContainerMenu> {
 
-    private final TauMenu menu;
+    private final UIMenu menu;
     private final List<Renderable> components;
     private final List<Renderable> tooltips;
     private final List<DynamicUIComponent> dynamicUIComponents;
@@ -32,7 +31,7 @@ public class TauContainerUI extends AbstractContainerScreen<TauContainerMenu> im
     private int left;
     private int top;
 
-    public TauContainerUI(TauContainerMenu pMenu, Inventory pPlayerInventory, TauMenu menu, boolean renderBackground, Theme theme) {
+    public TauContainerUI(TauContainerMenu pMenu, Inventory pPlayerInventory, UIMenu menu, boolean renderBackground, Theme theme) {
         super(pMenu, pPlayerInventory, Component.empty());
         this.menu = menu;
         this.renderBackground = renderBackground;
@@ -52,7 +51,7 @@ public class TauContainerUI extends AbstractContainerScreen<TauContainerMenu> im
         tooltips.clear();
         dynamicUIComponents.clear();
 
-        UIBuilder.build(layout, theme, menu.build(layout, theme), new BuildContext(components, tooltips, dynamicUIComponents, (List<GuiEventListener>) children(), new ArrayList<>()));
+        UIBuilder.build(layout, theme, menu.build(layout, theme, getMenu()), new BuildContext(components, tooltips, dynamicUIComponents, (List<GuiEventListener>) children(), new ArrayList<>()));
 
         built = true;
     }