+++ src/main/java/com/github/wintersteve25/tau/menu/UIMenu.java
@@ -8,21 +8,21 @@ import com.github.wintersteve25.tau.theme.MinecraftTheme;
 import com.github.wintersteve25.tau.theme.Theme;
 import com.github.wintersteve25.tau.utils.SimpleVec2i;
 import net.minecraft.client.gui.screens.MenuScreens;
+import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
 import net.minecraft.world.entity.player.Inventory;
 import net.minecraft.world.flag.FeatureFlagSet;
 import net.minecraft.world.inventory.MenuType;
 import net.neoforged.neoforge.client.event.RegisterMenuScreensEvent;
-import net.neoforged.neoforge.items.IItemHandler;
-import net.neoforged.neoforge.items.SlotItemHandler;
+import net.neoforged.neoforge.common.extensions.IMenuTypeExtension;
 import net.neoforged.neoforge.registries.DeferredHolder;
 import net.neoforged.neoforge.registries.DeferredRegister;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public interface TauMenu {
-    UIComponent build(Layout layout, Theme theme);
+public interface UIMenu {
+    UIComponent build(Layout layout, Theme theme, TauContainerMenu containerMenu);
 
     SimpleVec2i getSize();
 
@@ -35,33 +35,31 @@ public interface TauMenu {
     }
 
     default DeferredHolder<MenuType<?>, MenuType<TauContainerMenu>> registerMenuType(DeferredRegister<MenuType<?>> register, TauMenuHolder menu, String name, FeatureFlagSet featureFlagSet) {
-        return register.register(name, () -> new MenuType<>((cid, player) -> newMenu(menu, player, cid), featureFlagSet));
+        return register.register(name, () -> IMenuTypeExtension.create((cid, inv, data) -> newMenu(menu, inv, cid, data.readBlockPos())));
     }
 
     default void registerScreen(RegisterMenuScreensEvent event, MenuType<TauContainerMenu> menuType) {
         event.register(menuType, new MenuScreens.ScreenConstructor<TauContainerMenu, TauContainerUI>() {
             @Override
             public TauContainerUI create(TauContainerMenu pMenu, Inventory pInventory, Component pTitle) {
-                return new TauContainerUI(pMenu, pInventory, TauMenu.this, TauMenu.this.shouldRenderBackground(), TauMenu.this.getTheme());
+                return new TauContainerUI(pMenu, pInventory, UIMenu.this, UIMenu.this.shouldRenderBackground(), UIMenu.this.getTheme());
             }
         });
     }
 
-    default TauContainerMenu newMenu(TauMenuHolder menuHolder, Inventory playerInv, int containerId) {
+    default TauContainerMenu newMenu(TauMenuHolder menuHolder, Inventory playerInv, int containerId, BlockPos pos) {
         SimpleVec2i size = getSize();
         Layout layout = new Layout(size.x, size.y);
         List<MenuSlot<?>> s = new ArrayList<>();
 
-        UIComponent uiComponent = this.build(layout, getTheme());
+        TauContainerMenu menu = new TauContainerMenu(menuHolder.get(), playerInv, containerId, pos);
+        UIComponent uiComponent = this.build(layout, getTheme(), menu);
         UIBuilder.build(layout, getTheme(), uiComponent, new BuildContext(new ArrayList<>(), new ArrayList<>(), new ArrayList<>(), new ArrayList<>(), s));
 
-        return new TauContainerMenu(menuHolder.get(), playerInv, containerId) {
-            @Override
-            public void setupSlots(Inventory playerInventory) {
-                for (MenuSlot<?> slot : s) {
-                    slot.handler().setupSync(this, playerInventory, slot.pos().x, slot.pos().y);
-                }
-            }
-        };
+        for (MenuSlot<?> slot : s) {
+            slot.handler().setupSync(menu, playerInv, slot.pos().x, slot.pos().y);
+        }
+
+        return menu;
     }
 }